rename libsdl_xxx to libsdl2_xxx
